// src/payload-types.ts
export interface Media {
  id: string;
  alt?: string | null;
  caption?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    tablet?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface WowtourGalleryAlbumBlock {
  /**
   * Select background color for this section. Leaves blank for default.
   */
  backgroundColor?:
    | (
        | 'background'
        | 'foreground'
        | 'card'
        | 'card-foreground'
        | 'popover'
        | 'popover-foreground'
        | 'primary'
        | 'primary-foreground'
        | 'secondary'
        | 'secondary-foreground'
        | 'muted'
        | 'muted-foreground'
        | 'accent'
        | 'accent-foreground'
        | 'destructive'
        | 'destructive-foreground'
        | 'border'
        | 'input'
        | 'ring-3'
        | 'success'
        | 'warning'
        | 'error'
        | 'chart-1'
        | 'chart-2'
        | 'chart-3'
        | 'chart-4'
        | 'chart-5'
        | 'muted2'
        | 'muted2-foreground'
        | 'transparent'
      )
    | null;
  designVersion:
    | 'WOWTOUR_GALLERY_ALBUM_1'
    | 'WOWTOUR_GALLERY_ALBUM_2'
    | 'WOWTOUR_GALLERY_ALBUM_3'
    | 'WOWTOUR_GALLERY_ALBUM_4'
    | 'WOWTOUR_GALLERY_ALBUM_5';
  headingSettings?: {
    /**
     * block
     */
    heading?: string | null;
    /**
     * / Icon Heading
     */
    showHeadingIcon?: boolean | null;
    /**
     * SVG
     */
    headingIcon?: (string | null) | Media;
    /**
     * / Description Heading
     */
    showDescription?: boolean | null;
    /**
     * Heading
     */
    description?: string | null;
    /**
     * Tagline Heading ( "The journey is the destination")
     */
    tagline?: string | null;
  };
  albums?: (string | GalleryAlbum)[] | null;
  hiddenAlbumIds?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  cardSettings?: {
    /**
     * (default: 16px)
     */
    borderRadius?: number | null;
  };
  /**
   * 1 (default: 5)
   */
  albumsPerRow?: ('3' | '4' | '5' | '6') | null;
  /**
   * Grid Slide
   */
  displayMode?: ('showAll' | 'slide') | null;
  buttonSettings?: {
    /**
     * / ""
     */
    showButton?: boolean | null;
    /**
     * Button text
     */
    buttonText?: string | null;
    /**
     * (default: /gallery)
     */
    buttonLink?: string | null;
  };
  sliderSettings?: {
    /**
     * Auto play delay in milliseconds
     */
    autoPlayDelay?: number | null;
    /**
     * Enable/disable auto play
     */
    autoPlay?: boolean | null;
    /**
     * Enable/disable loop
     */
    loop?: boolean | null;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: 'wowtourGalleryAlbum';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "gallery-albums".
 */
export interface GalleryAlbum {
  id: string;
  /**
   * Album name (e.g., London, Paris, Tokyo)
   */
  title: string;
  slug?: string | null;
  slugLock?: boolean | null;
  /**
   * Hide this album from website (data not deleted)
   */
  isHidden?: boolean | null;
  /**
   * Album cover image (recommended 3:4 portrait ratio)
   */
  coverImage: string | Media;
  /**
   * All images in album
   */
  images?:
    | {
        image: string | Media;
        /**
         * Image caption (optional)
         */
        caption?: string | null;
        id?: string | null;
      }[]
    | null;
  meta?: {
    title?: string | null;
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (string | null) | Media;
    description?: string | null;
  };
  updatedAt: string;
  createdAt: string;
}